backend/requirements.txt
asgiref==3.8.1
astroid==3.3.9
click==8.1.8
dill==0.3.9
dj-database-url==2.3.0
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
h11==0.14.0
isort==6.0.1
mccabe==0.7.0
packaging==24.2
platformdirs==4.3.7
psycopg2-binary==2.9.10
PyJWT==2.9.0
pylint==3.3.6
python-dotenv==1.0.1
pytz==2025.1
sqlparse==0.5.3
tomlkit==0.13.2
typing_extensions==4.13.1
uvicorn==0.34.0
whitenoise==6.9.0
boto3==1.28.0
django-storages==1.13.2